Подключен SpringBoot Security

+++ src/main/java/by/sergey/cinemaservicespring/repository/ActorRepository.java
@@ -4,11 +4,10 @@ import by.sergey.cinemaservicespring.entity.Actor;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-
 
 @Repository
-public interface ActorRepository extends JpaRepository<Actor, Integer> {
+public interface ActorRepository extends JpaRepository<Actor, Long> {
+
     Actor getReferenceById(Long id);
-//    Actor getActorById(Long id);
+
 }
